+++ gluecodium/src/main/java/com/here/gluecodium/generator/dart/DartNameResolver.kt
@@ -103,8 +103,8 @@ internal class DartNameResolver(
             TypeId.STRING -> "String"
             TypeId.BLOB -> "Uint8List"
             TypeId.DATE -> "DateTime"
+            TypeId.DURATION -> "Duration"
             TypeId.LOCALE -> "Locale"
-            else -> "" // TODO: #911 Duration types
         }
 
     private fun resolveValue(limeValue: LimeValue): String =